refactor: reorganize code structure and improve documentation - Move commands/ to workflows/ for better semantic clarity - Maintain helpers/ for pure utility functions - Reserve commands/ namespace for future CLI commands - Update comprehensive README with usage examples and features
