+++ packages/create-fastify-project/src/helpers/error-handling.ts
