+++ packages/create-fastify-project/src/helpers/types.ts
@@ -1,0 +1,7 @@
+export interface ProjectOptions {
+  install: boolean;
+  git: boolean;
+  orm: "prisma" | "none";
+  db: "sqlite" | "postgres" | "mysql";
+  lint: "biome" | "eslint";
+}
