+++ packages/create-fastify-project/src/helpers/version-fetcher.ts
